pyproject.toml
[tool.poetry]
name = "swallow"
version = "0.1.0"
description = ""
authors = ["Dmitriy Shikhalev <dmitriy.shikhalev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
pyyaml = "^6.0.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0.25"}
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mypy = "^0.10.3"
pytest-pylint = "^0.19.0"
pytest-cov = "^4.0.0"
freezegun = "^1.2.2"
pika-stubs = "^0.1.3"
types-pyyaml = "^6.0.12.12"
pylint = "<3.0.0"
#pylint = "2.13.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
swallow = "swallow.main:main"
